refactor(clones): split god-module query_api/clones.rs into cohesive siblings
Decompose the 3347-line query_api/clones.rs into a directory module whose
mod.rs is a curated index over job-focused siblings:

- types.rs      public DTO / options / eligibility types
- substrate.rs  SymbolBag/TokenPosting + SourcererCC candidate-gen
- scoring.rs    ROI gates, dampen, apply_refinement, canonical-order helpers
- build.rs      build_class + staleness signal
- refine_load.rs  refine-input row hydration
- resolve.rs    selector resolution + ineligibility classification
- tests.rs      the unit tests, moved with the code they exercise

mod.rs keeps the IndexDatabase query methods on the type (so every call
site is unchanged) and re-exports the same pub surface. precompute.rs and
of_text.rs now import the candidate-gen primitives from substrate.

No logic change; no re-export shim. All call sites migrated to the new paths.
